Git processors improvements: - Refactor to organize all git related processors in a package - Git diff processor now makes the latest commit id available - Add new processor to only update the latest commit id
