+++ src/main/java/org/craftercms/deployer/impl/DeploymentConstants.java
@@ -51,6 +51,7 @@ public abstract class DeploymentConstants {
     // Processor params
 
     public static final String REPROCESS_ALL_FILES_PARAM_NAME = "reprocess_all_files";
+    public static final String LATEST_COMMIT_ID_PARAM_NAME = "latest_commit_id";
 
     // Logging MDC Keys
