+++ src/main/java/org/craftercms/deployer/impl/processors/git/GitPushProcessor.java
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.craftercms.deployer.impl.processors;
+package org.craftercms.deployer.impl.processors.git;
 
 import java.io.File;
 import java.util.ArrayList;
